+++ lib/prop_background.ml
@@ -225,7 +225,8 @@ let rec pp_border_radius : border_radius Pp.t =
 let normalize_border_radius ?(strip = true) : border_radius -> border_radius =
  fun value ->
   let group =
-    normalize_box_shorthand (Values.normalize_length_percentage ~strip)
+    normalize_box_shorthand ~is_substitution:is_lp_substitution
+      (Values.normalize_length_percentage ~strip)
   in
   match value with
   | Radius { horizontal; vertical } ->
@@ -233,8 +234,10 @@ let normalize_border_radius ?(strip = true) : border_radius -> border_radius =
       let vertical = option_map_preserve group vertical in
       let vertical =
         match vertical with
-        | Some vs when List.equal Values.equal_length_percentage horizontal vs
-          ->
+        | Some vs
+          when (not (List.exists is_lp_substitution horizontal))
+               && (not (List.exists is_lp_substitution vs))
+               && List.equal Values.equal_length_percentage horizontal vs ->
             Option.None
         | _ -> vertical
       in